Added support for Realm.removeListener
Fixed RealmCoordinator callback comparison in native code used removeListener is called
Add API docs to Realm class
changed Realm.exists and Realm.schemaVersion to take path to realm file.

+++ lib/src/configuration.dart
@@ -167,7 +167,7 @@ class Configuration {
   // set List<Type> schema_objects;
 
   SchemaObjects schema;
-  int schemaVersion;
+  double schemaVersion;
   bool deleteRealmIfMigrationNeeded;
   bool disableFormatUpgrade;
